backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt
passlib[bcrypt]
python-multipart
openpyxl
reportlab
python-dotenv
pytz
twilio
alembic